@@ -9,7 +9,7 @@
         }
     ],
     "require": {
-		"php":">=5.3.0",
+		"php":">=7.1",
 	    "ext-curl":"*"
     },
 	"autoload":{
